Be more strict about when inventory units and shipments are allowed to be destroyed.
originally from https://github.com/bonobos/spree/pull/406

+++ core/app/models/spree/order.rb
@@ -3,6 +3,7 @@ require 'spree/order/checkout'
 
 module Spree
   class Order < Spree::Base
+    class CannotRebuildShipments < StandardError; end
 
     ORDER_NUMBER_LENGTH  = 9
     ORDER_NUMBER_LETTERS = false
@@ -522,9 +523,14 @@ module Spree
     end
 
     def create_proposed_shipments
-      adjustments.shipping.delete_all
-      shipments.destroy_all
-      self.shipments = Spree::Stock::Coordinator.new(self).shipments
+      if completed?
+        raise CannotRebuildShipments.new(Spree.t(:cannot_rebuild_shipments_order_completed))
+      elsif shipments.any? { |s| !s.pending? }
+        raise CannotRebuildShipments.new(Spree.t(:cannot_rebuild_shipments_shipments_not_pending))
+      else
+        shipments.destroy_all
+        self.shipments = Spree::Stock::Coordinator.new(self).shipments
+      end
     end
 
     def apply_free_shipping_promotions
@@ -540,7 +546,7 @@ module Spree
     # to delivery again so that proper updated shipments are created.
     # e.g. customer goes back from payment step and changes order items
     def ensure_updated_shipments
-      unless completed?
+      if !completed? && shipments.all?(&:pending?)
         self.shipments.destroy_all
         self.update_column(:shipment_total, 0)
         restart_checkout_flow
